% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpfiles.r
\name{AudioTest}
\alias{AudioTest}
\title{Test and play audio in learnr package}
\description{
Test play sound
}
\examples{
\dontrun{
learnr::run_tutorial("audioTest", package = "ecoSoundr")
}
}
\keyword{ecological}
\keyword{sounds}
